requirements.txt
torch~=1.8.1
transformers==4.4.2
tqdm==4.59.0
h5py==3.2.1
numpy~=1.20.1
torchvision~=0.9.1
Pillow==8.1.2
fire==0.4.0
nltk==3.6.2
gensim==4.0.1
bert_score==0.3.9
mosestokenizer==1.1.0
sentencepiece==0.1.96
